feat: float32 vocoder, 48kHz output, gapless streaming playback
- Cast vocoder output to float32 before waveform reconstruction for
  reduced numerical noise
- Add sample_rate parameter (24000 or 48000) across the public API with
  numpy-only FFT resampling for the 2x case
- Rewrite play_stream with a persistent OutputStream and producer thread
  to eliminate inter-chunk gaps

+++ src/kokoro_mlx/istftnet.py
@@ -495,7 +495,7 @@ class Generator(nn.Module):
             x = xs / self.num_kernels
 
         x = nn.leaky_relu(x)
-        x = self.conv_post(x)  # (B, n_fft+2, T)
+        x = self.conv_post(x).astype(mx.float32)  # float32 for precise waveform reconstruction
         spec = mx.exp(x[:, :self.post_n_fft // 2 + 1, :])
         phase = mx.sin(x[:, self.post_n_fft // 2 + 1:, :])
         return self.stft.inverse(spec, phase)  # (B, 1, samples)
